+++ src/hoppet.h
@@ -228,22 +228,21 @@ extern "C" {
   void hoppetAssign(void (* pdf_subroutine)(const double & x, 
                                             const double & Q, double * res) );
 
-//=====================================================================
-// Set up the strong coupling such that alphas(Q)=alphas_Q, with the
-// given number of loops (nloop).
-//
-// The user should have set the quark masses or requested a FFN scheme
-// prior to calling this function.
-//
-// This function is provided mainly for use in conjunction with hoppetAssign.
-// In particular, it has the side effect of modifying the structure of the
-// PDF tables to make sure they know about the mass thresholds.
-//
-// If QED has been requested, a QED coupling will also be set up
-// (its value is not currently configurable from this interface).
-//
-// If you call hoppetEvolve (below), there is no need to separately call
-// hoppetSetCoupling.
+  /// Set up the strong coupling such that alphas(Q)=alphas_Q, with the
+  /// given number of loops (nloop).
+  ///
+  /// The user should have set the quark masses or requested a FFN scheme
+  /// prior to calling this function.
+  ///
+  /// This function is provided mainly for use in conjunction with hoppetAssign.
+  /// In particular, it has the side effect of modifying the structure of the
+  /// PDF tables to make sure they know about the mass thresholds.
+  ///
+  /// If QED has been requested, a QED coupling will also be set up
+  /// (its value is not currently configurable from this interface).
+  ///
+  /// If you call hoppetEvolve (below), there is no need to separately call
+  /// hoppetSetCoupling.
   void hoppetSetCoupling(const double & asQ0,
                          const double & Q0alphas,
                          const int    & nloop);
@@ -252,6 +251,13 @@ extern "C" {
   /// Given a pdf_subroutine with the interface shown below, fill the 
   /// table by evolving the PDF from scale Q0pdf, with alphas provided 
   /// at scale Q0alphas
+  ///
+  /// @param Q0alphas       Initial scale for alpha_s [GeV]
+  /// @param nloop          Number of loops for evolution (1=LO, 2=NLO, 3=NNLO, 4=N3LO)
+  /// @param muR_Q          Ratio of renormalisation scheme to factorisation scale during evolution
+  /// @param pdf_subroutine Function pointer to the user-defined PDF function
+  /// @param Q0pdf          Initial scale for PDF [GeV]
+  ///
   void hoppetEvolve(const double & asQ0,
                     const double & Q0alphas,
                     const int    & nloop,
@@ -261,7 +267,16 @@ extern "C" {
                     const double & Q0pdf);
 
 
-  /// Prepare a cached evolution
+  /// Prepare a cached evolution.
+  /// Once this has been called, one can use hoppetCachedEvolve to
+  /// carry out the cached evolution of a specific initial condition.
+  ///
+  /// @param asQ0     Strong coupling at initial scale Q0alphas
+  /// @param Q0alphas Initial scale for alpha_s [GeV]
+  /// @param nloop    Number of loops for evolution (1=LO, 2=NLO, 3=NNLO, 4=N3LO)
+  /// @param muR_Q    Ratio of renormalisation scheme to factorisation scale during evolution
+  /// @param Q0pdf    Initial scale for PDF [GeV]
+  ///
   void hoppetPreEvolve(const double & asQ0, 
                        const double & Q0alphas, 
                        const int    & nloop,